allow plugins to effect route status
