+++ controller/pkg/agentgateway/plugins/registry.go
@@ -37,6 +37,10 @@ func MergePlugins(plug ...AgwPlugin) AgwPlugin {
 			if p.AddResourceExtension.AncestorBackends != nil {
 				ret.AddResourceExtension.AncestorBackends = p.AddResourceExtension.AncestorBackends
 			}
+			if p.AddResourceExtension.GatewayStatuses != nil {
+				ret.AddResourceExtension.GatewayStatuses = p.AddResourceExtension.GatewayStatuses
+			}
+			ret.AddResourceExtension.ParentResolvers = append(ret.AddResourceExtension.ParentResolvers, p.AddResourceExtension.ParentResolvers...)
 		}
 	}
 	return ret
